+++ typeDensityInterface.go
@@ -1,9 +1,9 @@
 package iotmaker_platform_coordinate
 
 type IDensity interface {
-	Set(value int)
-	Add(value int)
-	Sub(value int)
+	Set(value float64)
+	Add(value float64)
+	Sub(value float64)
 	SetDensityFactor(value interface{})
 	Int() int
 	Float64() float64
